requirements.txt
python-dotenv
discord.py
discord-py-slash-command
